Refactor GaussianNode
This commit modifies the parametrization of the GaussianNode.
The new parametrization is: GaussianNode{mean_type,uncertainty_type}
mean_type can be :mean or :fixed_mean
uncertainty_type can be: :variance, :fixed_variance, :precision,
:log_variance.

The new parametrization is useful for dispatching msg calc rules
based on both the node form and on which parameters are fixed.

Also added sum-product rules for fixed variance case.