Implement approximate message computation rule mechanism
Approximate rules are identified by an extra argument in the calling signature.
The argument specifies the type of Approximation (<: ApproximationType).
Algorithm constructors take keyword argument message_types to fix
messages to a specific type. message_types can also contain an
Approximation type, which specifies both the type of the approximating
distribution and the approximation method (such as MomentMatching).